requirements.txt
slack-sdk==3.26.1
datadog-api-client==2.21.0
python-dotenv==1.0.0
openai==1.12.0
fastapi==0.109.2
uvicorn==0.27.1 